src/py-receipt-server/requirements.txt
transformers[torch]==4.33.2
pillow
flask
numpy
